APPEDOLT_3.1.013
Major releease.

1. Parallel connection issue . APPEDOLT_3.1.012
The prallel threads where formed only when the number requests matches with the number of parallel connection
set in the UI. All the other request where gone in sequential.

This has been modified now such that, if the requstes are less than or more than parallel connection set values ,
The parallel threads are formed in a batch of set value from the UI .

Observed some time the missing request and response .
This has been resolved when parallel connection logic has been rewritten.

2. The RUN result, the hit count values are mismatching and in reports the values are missing for 100 user run
Observed the values are writing correctly in the logs kept after queue drains but missing in the SQLITE DB .
Changed the logic of DB writing from the Queue , multiple CSV file writing is implemented
with multithreading concept to drain and write into the DB .

Minor changes

3. RUN button not visible after script run.
   Observed sometime in the long run cases , the report run loop goes in infinite loop and never comesout .
   handled this situation.

4. Time is in miliseconds : The run value time is writing in DB in miliseconds, instead of seconds.

5. After parallel logic implementation , The variables value passed through parameterization is not reflecting
   during the run time, This has been rectified.